Adding possibility to switch between euro and USD

+++ app/src/main/java/com/openclassrooms/realestatemanager/utils/Utils.kt
@@ -19,7 +19,7 @@ object Utils {
      * @param dollars
      * @return
      */
-    fun convertDollarToEuro(dollars: Int): Int {
+    fun convertDollarToEuro(dollars: Long): Int {
         return (dollars * 0.94345).roundToInt()
     }
